Nuked for_each
