+++ include/tiny_list.h
@@ -19,6 +19,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
+#include "tiny_utils.h"
 
 typedef struct tiny_list_node_t {
   struct tiny_list_node_t* next;
@@ -28,7 +29,11 @@ typedef struct {
   tiny_list_node_t head;
 } tiny_list_t;
 
-typedef bool (*tiny_list_for_each_t)(tiny_list_node_t* node, uint16_t index, void* context);
+typedef struct
+{
+  tiny_list_t* list;
+  tiny_list_node_t* current;
+} tiny_list_iterator_t;
 
 /*!
  * Initializes the list.
@@ -76,9 +81,23 @@ bool tiny_list_contains(tiny_list_t* self, tiny_list_node_t* node);
 uint16_t tiny_list_index_of(tiny_list_t* self, tiny_list_node_t* node);
 
 /*!
- * Invokes the provided callback for each node in the list.
- * Traverses from head to tail and stops iteration if the callback returns false.
+ * Initialize an iterator for the provided list.
  */
-void tiny_list_for_each(tiny_list_t* self, tiny_list_for_each_t callback, void* context);
+void tiny_list_iterator_init(tiny_list_iterator_t* self, tiny_list_t* list);
+
+/*!
+ * Return a pointer to the next node or NULL if there are no more nodes.
+ */
+tiny_list_node_t* tiny_list_iterator_next(tiny_list_iterator_t* self);
+
+#define tiny_list_iterate(_list, _type, _item, ...)          \
+  do {                                                       \
+    tiny_list_iterator_t _it;                                \
+    tiny_list_iterator_init(&_it, _list);                    \
+    _type* _item;                                            \
+    while((_item = (_type*)tiny_list_iterator_next(&_it))) { \
+      __VA_ARGS__                                            \
+    }                                                        \
+  } while(0)
 
 #endif
