+++ src/tiny_event.c
@@ -29,24 +29,9 @@ void tiny_event_init(tiny_event_t* self)
   tiny_list_init(&self->subscribers);
 }
 
-typedef struct {
-  const void* args;
-  tiny_event_t* self;
-} publish_context_t;
-
-static bool publish(tiny_list_node_t* node, uint16_t index, void* _context)
-{
-  (void)index;
-  reinterpret(context, _context, publish_context_t*);
-  reinterpret(subscription, node, tiny_event_subscription_t*);
-
-  tiny_event_subscription_publish(subscription, context->args);
-
-  return true;
-}
-
 void tiny_event_publish(tiny_event_t* self, const void* args)
 {
-  publish_context_t context = { args, self };
-  tiny_list_for_each(&self->subscribers, publish, &context);
+  tiny_list_iterate(&self->subscribers, tiny_event_subscription_t, subscription, {
+    tiny_event_subscription_publish(subscription, args);
+  });
 }
